dashboard/requirements.txt
streamlit==1.32.0
pandas==2.2.0
clickhouse-driver==0.2.6
plotly==5.18.0
python-dotenv==1.0.1 